refactor(hub): use queryparams to route validators
