+++ apps/hub/src/app/validators/validator/undelegate-content.tsx
@@ -17,7 +17,7 @@ import { useTheme } from "next-themes";
 import { Address, parseUnits } from "viem";
 
 import ValidatorInput from "~/components/validator-input";
-import { DelegateEnum, ImageMapEnum } from "../types";
+import { DelegateEnum, ImageMapEnum } from "./types";
 
 export const UnDelegateContent = ({
   userValidator,
