+++ apps/hub/src/app/validators/validator/validator-details.tsx
